sigma: generalized type checking for sigma
